* develop: (120 commits)
  Open default browser using npm start
  Put favicon on the right folder
  Hide the other options
  Run all tests or run one or more independantly
  Create automated tests of filters
  Refactor tests file
  Change titles from "Sort by"
  Create an automated tests file for inline editing
  Add export automated tests to branch 199
  Add cascade to clean time tracker schema correctly
  Clean database again
  Clean database
  Clean entries on database
  Use psql cmd instead of pg_restore
  Create a backup of  time tracker schema only
  Change database user
  Fix bug about "all" unselecting by itself
  Adapt all files to run automatically
  Adapt create edit items
  Creata a frozen dataset
  ...

# Conflicts:
#	package.json

+++ src/app/time-tracking/components/entries/delete-entry/delete-entry.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { MdDialogRef } from '@angular/material';
-import { TimeTrackingEntryService } from '../../../../data';
+import { TimeTrackingEntryService, RegistryService } from '../../../../data';
 import { Http } from '@angular/http';
 import { environment } from '../../../../../environments/environment';
+import { LoginService } from '../../../../login';
 
 @Component({
   selector: 'app-delete-entry',
@@ -18,14 +19,22 @@ export class DeleteEntryComponent implements OnInit {
 
   constructor(public dialogRef: MdDialogRef<DeleteEntryComponent>,
     public timeTrackingEntryService: TimeTrackingEntryService,
+    public loginService: LoginService,
+    public registryService: RegistryService,
     private http: Http) {
   }
 
   public deleteEntry() {
-    this.http.delete(this.baseUrl + "/timeentries/" + this.rowid).map(res => res.json()).subscribe(
-      () => {
-        this.dialogRef.close(true);
-      });
+    if (this.loginService.loggedIn()) {
+      this.http.delete(this.baseUrl + "/timeentries/" + this.rowid).map(res => res.json()).subscribe(
+        () => {
+          this.registryService.entriesComponent.offset = 0;
+          this.dialogRef.close(true);
+        });
+    } else {
+      alert("Your token has expired. Please log in again!");
+      this.dialogRef.close(true);
+    }
   }
 
   ngOnInit() {
